CRAYSAT-1411: Refactor image dependencies and add name templating
Add the ability to use the name of the base image or recipe that was
used to construct an image as a variable when templating the `name`
property of an image.

Add the ability to use the name of the image as a variable when
templating the `name` property of a session template.

Since names of session templates and images are now templated and not
required to be unique and cannot be rendered until the base is
validated, change the `__str__` methods of both `InputSessionTemplate`
and `BaseInputImage` to simply state the index of the item in the input
instance.

Also add general variable substitution (e.g. product versions) to the
`conifguration` property for images and to the `name`, `image`, and
`configuration` properties of session templates.

Since image names are now templated using variables from the base of the
image, it is necessary to change how images in a `sat bootprep` input
file refer to other images as a base. Add a `ref_name` property on the
images which gives them a unique static name that can be referenced by
another image's `base.image_ref` property. It is no longer supported to
specify a dependency between images using the IMS image name.

Now that we are no longer checking for dependencies between images based
on the images' IMS names, add a check to ensure there is no overlap
between the IMS names of images created by the input file and the IMS
names of images used as bases. This would result in possible race
conditions and non-deterministic behavior if the base of on image is
created by another image in the input file.

Test Description:
Tested on vale with a bootprep input file that builds an image from the
recipe provided by the CSM product, and then customizes it with a second
image specification that refers to the first by its `ref_name`.

Also tested using the `image.name` variable substitution in BOS session
template names.