CRAYSAT-1411: Do `find_image_dependencies` on all images
If we call `find_image_dependencies` only on the list of images returned
by `handle_existing_images`, we can end up with unresolvable
dependencies if they already exist, and the user decided to skip them.
Instead, call `find_image_dependencies` on the full list of
`input_images` from the instance.

Test Description:
Tested this change on vale with a situation where image A depended on B
and A was skipped because it already existed.