CRAYSAT-1410: Address validate_image_exists bug
This commit addresses an edge case in which an image in the bootprep
configuration is referred to from a session template in that
configuration. In this case, the session template should validate even
when the image does not yet exist, because the image will be created
as a part of that `sat bootprep` run.

This commit also removes two unused imports from session_template.py.

Test Description:
Ran `sat bootprep` in dry-run mode using two different configurations:
* A configuration that builds an image, then creates a session template
  that uses that image, using the "old"-style format of referring
  directly to the image by name.
* A configuration that also builds an image and session template with
  that image, but using the "new"-style format of referring to the image
  using `ref_name` and `image_ref`.